AGED-158: Match message feature to firestore rules

+++ news/domain/src/main/java/com/upsaclay/news/domain/usecase/DeleteAnnouncementUseCase.kt
@@ -1,13 +1,20 @@
 package com.upsaclay.news.domain.usecase
 
+import com.upsaclay.common.domain.ConnectivityObserver
+import com.upsaclay.common.domain.entity.NoInternetConnectionException
 import com.upsaclay.news.domain.entity.Announcement
 import com.upsaclay.news.domain.entity.AnnouncementState
 import com.upsaclay.news.domain.repository.AnnouncementRepository
 
 class DeleteAnnouncementUseCase(
-    private val announcementRepository: AnnouncementRepository
+    private val announcementRepository: AnnouncementRepository,
+    private val connectivityObserver: ConnectivityObserver
 ) {
     suspend operator fun invoke(announcement: Announcement) {
+        if (!connectivityObserver.isConnected) {
+            throw NoInternetConnectionException()
+        }
+
         if (announcement.state == AnnouncementState.PUBLISHED) {
             announcementRepository.deleteAnnouncement(announcement)
         } else {
